requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
pytorch
scikit-learn
scipy
transformers
